Clean up of VS2015 warnings - Part 1 review changes
Fix items from review

+++ native/dll/DShowCapture-2/ScanChannel.cpp
@@ -139,7 +139,7 @@ JNIEXPORT jstring JNICALL Java_sage_DShowCaptureDevice_scanBDAChannel0
 			   
 	char* scanResult = scanChannel( (CHANNEL_DATA*)pCapInfo->channel, deviceName, tuningString,  
 		                               countryRegion,  (int)streamType );
-	(size_t) len = strlen( scanResult );
+	len = strlen( scanResult );
 	//trim \r\n
 	while ( len > 1 && scanResult[len-1] == '\r' || scanResult[len-1] == '\n' ) 
 	{
